Squashed commit of the following:
commit 7df03864ff903a00022d1eeb891a7be4a28a0015
Author: MirandaWood <miranda.Wood@uk.ey.com>
Date:   Tue Feb 22 15:25:33 2022 +0000

    feat: moved mappingkey to sep class, fixes in increments visitor
