+++ src/traverse/Indicator.ts
@@ -2,6 +2,7 @@
 
 import NodePath from './NodePath.js';
 import Scope from './Scope.js';
+import MappingKey from './MappingKey.js';
 import { VariableBinding } from './Binding.js';
 import logger from '../utils/logger.js';
 import backtrace from '../error/backtrace.js';
@@ -609,191 +610,5 @@ export class StateVariableIndicator extends FunctionDefinitionIndicator {
       }
     });
     this.newCommitmentsRequired = !burnedOnly;
-  }
-}
-
-/**
- * If a Binding/StateVarIndicator represents a mapping, it will contain a MappingKey class.
- */
-export class MappingKey extends StateVariableIndicator {
-
-  container: any;
-  referencedKeyId: number;
-  referencedKeyNodeType: string;
-  referencedKeyIsParam: boolean; // is a function parameter - used for finding owner
-  referencedKeyName: string;
-  keyPath: NodePath;
-  isMsgSender: boolean; // used for finding owner
-  /**
-   * A mappingKey can be contained within a binding or an indicator class.
-   * @param { Binding || StateVarIndicator } container
-   * @param { NodePath } keyPath
-   */
-  constructor(container: any, keyPath: NodePath) {
-    super(keyPath)
-    this.container = container;
-    this.id = container.id;
-    this.node = container.node;
-
-    // TODO: distinguish between if the key is a reference and if the key is not a reference - the prefix 'referenced' is misleading below:
-    this.referencedKeyId = keyPath.node.referencedDeclaration;
-    this.referencedKeyName = keyPath.isMsg()
-      ? 'msg'
-      : keyPath.getReferencedNode().name;
-    this.referencedKeyNodeType = keyPath.isMsg()
-      ? 'msg.sender'
-      : keyPath.getReferencedNode().nodeType;
-    this.referencedKeyIsParam = keyPath.isFunctionParameter(); // is a function parameter - used for finding owner
-    this.isMsgSender = keyPath.isMsg(); // used for finding owner
-    this.isSecret = container.isSecret; // only really used by binding.
-    this.isMapping = true;
-
-    this.name = this.isMsgSender
-      ? `${container.name}[msg.sender]`
-      : `${container.name}[${keyPath.node.name}]`;
-
-  }
-
-  updateProperties(path: NodePath) {
-    this.addReferencingPath(path);
-    this.isUnknown ??= path.node.isUnknown;
-    this.isKnown ??= path.node.isKnown;
-    this.reinitialisable ??= path.node.reinitialisable;
-    if (path.isModification()) this.addModifyingPath(path);
-
-    this.container.updateProperties(path);
-  }
-
-  // TODO: move into commonFunctions (because it's the same function as included in the Indicator class)
-  addReferencingPath(path: NodePath) {
-    this.isReferenced = true;
-    ++this.referenceCount;
-    if (!this.referencingPaths.some(p => p.node.id === path.node.id))
-      this.referencingPaths.push(path);
-  }
-
-  addModifyingPath(path: NodePath) {
-    this.isModified = true;
-    ++this.modificationCount;
-    if (!this.modifyingPaths.some(p => p.node.id === path.node.id)) {
-      this.modifyingPaths.push(path);
-
-      this.initialisationRequired = true; // Used? Probably for whole states?
-
-      const { node } = path;
-      if (node.isKnown) this.isKnown = true;
-      if (node.isUnknown) this.isUnknown = true;
-    }
-  }
-
-  addNullifyingPath(path: NodePath) {
-    this.isNullified = true;
-    ++this.nullificationCount;
-    this.nullifyingPaths.push(path);
-  }
-
-  addBurningPath(path: NodePath) {
-    this.isBurned = true;
-    this.burningPaths.push(path);
-  }
-
-  addSecretInteractingPath(path: NodePath) {
-    this.interactsWithSecret = true;
-    path.isSecret = true;
-    if (!this.interactsWith.some(p => p.node.id === path.node.id)) {
-      this.interactsWith.push(path);
-      this.container.addSecretInteractingPath(path);
-    }
-  }
-
-  addPublicInteractingPath(path: NodePath) {
-    this.interactsWithPublic = true;
-    path.isPublic = true;
-    if (!this.interactsWith.some(p => p.node.id === path.node.id)) {
-      this.interactsWith.push(path);
-      this.container.addPublicInteractingPath(path);
-    }
-  }
-
-  prelimTraversalErrorChecks() {
-    // warning: state is clearly whole, don't need known decorator
-    if (this.isKnown && this.isWhole) {
-      logger.warn(
-        `PEDANTIC: Unnecessary 'known' decorator. Secret state '${this.name}' is trivially 'known' because it is 'whole', due to: ${this.isWholeReason}`,
-      );
-      this.isWholeReason?.forEach(reason => {
-        console.log(reason[0]);
-      });
-    }
-    // error: conflicting unknown/whole state
-    if (this.isUnknown && this.isWhole) {
-      throw new SyntaxUsageError(
-        `Can't mark a whole state as 'unknown'`,
-        this.node,
-        this.isWholeReason,
-      );
-    }
-    // error: conflicting whole/partitioned state
-    if (this.isWhole && this.isPartitioned) {
-      throw new SyntaxUsageError(
-        `State cannot be whole and partitioned. The following reasons conflict.`,
-        this.container.node,
-        [...this.isWholeReason, ...this.isPartitionedReason],
-      );
-    }
-  }
-
-  updateIncrementation(path: NodePath, state: any) {
-    if (!path.isIncremented || state.incrementedIdentifier.isKnown) {
-      this.isWhole = true;
-      const reason = { src: state.incrementedIdentifier.src, 0: `Overwritten` };
-      this.isWholeReason ??= [];
-      this.isWholeReason.push(reason);
-      if (state.incrementedPath && !state.incrementedIdentifier.reinitialisable)
-        this.addNullifyingPath(state.incrementedPath);
-    } else if (
-      !path.isDecremented &&
-      (state.incrementedIdentifier.isUnknown ||
-        state.incrementedIdentifier.baseExpression?.isUnknown)
-    ) {
-      this.isPartitioned = true;
-      const reason = {
-        src: state.incrementedIdentifier.src,
-        0: `Incremented and marked as unknown`,
-      };
-      this.isUnknown ??= true;
-      this.isPartitionedReason ??= [];
-      this.isPartitionedReason.push(reason);
-    }
-    if (path.isDecremented && !state.incrementedIdentifier.isKnown)
-      this.addNullifyingPath(state.incrementedPath);
-    // if its incremented anywhere, isIncremented = true
-    // so we only assign if it's already falsey
-    this.isIncremented ||= path.isIncremented;
-    this.isDecremented ||= path.isDecremented;
-    this.increments ??= [];
-    this.decrements ??= [];
-    state.increments.forEach((inc: any) => {
-      this.increments.push(inc);
-    });
-    state.decrements.forEach((dec: any) => {
-      this.decrements.push(dec);
-    });
-  }
-
-  updateFromBinding() {
-    // it's possible we dont know in this fn scope whether a state is whole/owned or not, but the binding (contract scope) will
-    this.isWhole ??= this.container.binding.isWhole;
-    this.isWholeReason = this.isWhole
-      ? this.container.binding.isWholeReason
-      : this.isWholeReason;
-    this.isPartitioned ??= this.container.binding.isPartitioned;
-    this.isPartitionedReason = this.isPartitioned
-      ? this.container.binding.isPartitionedReason
-      : this.isPartitionedReason;
-    this.isOwned ??= this.container.binding.isOwned;
-    this.owner ??= this.container.binding.owner;
-    this.mappingOwnershipType = this.owner?.mappingOwnershipType;
-    this.onChainKeyRegistry ??= this.container.binding.onChainKeyRegistry;
   }
 }
